refactor

+++ src/main/java/tgb/btc/rce/service/handler/impl/state/BanUnbanStateHandler.java
@@ -1,24 +1,27 @@
-package tgb.btc.rce.service.handler.impl.message.text.state;
+package tgb.btc.rce.service.handler.impl.state;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.BooleanUtils;
 import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.objects.Message;
+import org.telegram.telegrambots.meta.api.objects.Update;
 import tgb.btc.library.interfaces.service.bean.bot.user.IReadUserService;
 import tgb.btc.library.service.process.BannedUserCache;
 import tgb.btc.library.service.process.BanningUserService;
 import tgb.btc.rce.enums.BotReplyButton;
 import tgb.btc.rce.enums.Menu;
 import tgb.btc.rce.enums.UserState;
+import tgb.btc.rce.enums.update.TextCommand;
+import tgb.btc.rce.enums.update.UpdateType;
 import tgb.btc.rce.sender.IResponseSender;
 import tgb.btc.rce.service.IRedisUserStateService;
-import tgb.btc.rce.service.handler.message.text.IStateTextMessageHandler;
+import tgb.btc.rce.service.handler.IStateHandler;
 import tgb.btc.rce.service.handler.util.IAdminPanelService;
 import tgb.btc.rce.service.util.IMenuService;
 
 @Service
 @Slf4j
-public class BanUnbanStateHandler implements IStateTextMessageHandler {
+public class BanUnbanStateHandler implements IStateHandler {
 
     private final IReadUserService readUserService;
 
@@ -48,7 +51,13 @@ public class BanUnbanStateHandler implements IStateTextMessageHandler {
     }
 
     @Override
-    public void handle(Message message) {
+    public void handle(Update update) {
+        if (!update.hasMessage() || !update.getMessage().hasText()) {
+            responseSender.sendMessage(UpdateType.getChatId(update),
+                    "Введите chat id пользователя, либо нажмите \"" + TextCommand.CANCEL.getText() + "\".");
+            return;
+        }
+        Message message = update.getMessage();
         Long chatId = message.getChatId();
         if (message.getText().equals(BotReplyButton.CANCEL.getText())) {
             redisUserStateService.delete(chatId);
