+++ src/main/java/tgb/btc/rce/service/handler/impl/state/DealsCountStateHandler.java
@@ -1,17 +1,20 @@
-package tgb.btc.rce.service.handler.impl.message.text.state;
+package tgb.btc.rce.service.handler.impl.state;
 
 import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.objects.Message;
+import org.telegram.telegrambots.meta.api.objects.Update;
 import tgb.btc.library.service.properties.FunctionsPropertiesReader;
 import tgb.btc.rce.enums.BotReplyButton;
 import tgb.btc.rce.enums.UserState;
+import tgb.btc.rce.enums.update.TextCommand;
+import tgb.btc.rce.enums.update.UpdateType;
 import tgb.btc.rce.sender.IResponseSender;
 import tgb.btc.rce.service.IRedisUserStateService;
-import tgb.btc.rce.service.handler.message.text.IStateTextMessageHandler;
+import tgb.btc.rce.service.handler.IStateHandler;
 import tgb.btc.rce.service.handler.util.IAdminPanelService;
 
 @Service
-public class DealsCountStateHandler implements IStateTextMessageHandler {
+public class DealsCountStateHandler implements IStateHandler {
 
     private final FunctionsPropertiesReader functionsPropertiesReader;
 
@@ -31,7 +34,13 @@ public class DealsCountStateHandler implements IStateTextMessageHandler {
     }
 
     @Override
-    public void handle(Message message) {
+    public void handle(Update update) {
+        if (!update.hasMessage() || !update.getMessage().hasText()) {
+            responseSender.sendMessage(UpdateType.getChatId(update),
+                    "Введите новое количество возможных активных сделок, либо нажмите \"" + TextCommand.CANCEL.getText() + "\".");
+            return;
+        }
+        Message message = update.getMessage();
         Long chatId = message.getChatId();
         if (message.getText().equals(BotReplyButton.CANCEL.getText())) {
             redisUserStateService.delete(chatId);
