+++ src/main/java/tgb/btc/rce/service/handler/impl/state/RankDiscountStateHandler.java
@@ -1,16 +1,19 @@
-package tgb.btc.rce.service.handler.impl.message.text.state;
+package tgb.btc.rce.service.handler.impl.state;
 
 import org.apache.commons.lang3.BooleanUtils;
 import org.springframework.stereotype.Service;
 import org.telegram.telegrambots.meta.api.objects.Message;
+import org.telegram.telegrambots.meta.api.objects.Update;
 import tgb.btc.library.interfaces.service.bean.bot.IUserDiscountService;
 import tgb.btc.library.interfaces.service.bean.bot.user.IReadUserService;
 import tgb.btc.rce.enums.BotReplyButton;
 import tgb.btc.rce.enums.UserState;
 import tgb.btc.rce.enums.update.CallbackQueryData;
+import tgb.btc.rce.enums.update.TextCommand;
+import tgb.btc.rce.enums.update.UpdateType;
 import tgb.btc.rce.sender.IResponseSender;
 import tgb.btc.rce.service.IRedisUserStateService;
-import tgb.btc.rce.service.handler.message.text.IStateTextMessageHandler;
+import tgb.btc.rce.service.handler.IStateHandler;
 import tgb.btc.rce.service.handler.util.IAdminPanelService;
 import tgb.btc.rce.service.impl.keyboard.KeyboardBuildService;
 import tgb.btc.rce.service.util.ICallbackDataService;
@@ -19,7 +22,7 @@ import tgb.btc.rce.vo.InlineButton;
 import java.util.List;
 
 @Service
-public class RankDiscountStateHandler implements IStateTextMessageHandler {
+public class RankDiscountStateHandler implements IStateHandler {
 
     private final IReadUserService readUserService;
 
@@ -49,7 +52,13 @@ public class RankDiscountStateHandler implements IStateTextMessageHandler {
     }
 
     @Override
-    public void handle(Message message) {
+    public void handle(Update update) {
+        if (!update.hasMessage() || !update.getMessage().hasText()) {
+            responseSender.sendMessage(UpdateType.getChatId(update),
+                    "Введите chat id пользователя, либо нажмите \"" + TextCommand.CANCEL.getText() + "\".");
+            return;
+        }
+        Message message = update.getMessage();
         Long chatId = message.getChatId();
         if (message.getText().equals(BotReplyButton.CANCEL.getText())) {
             redisUserStateService.delete(chatId);
